Add logging facility.
In more than one occasion we had to inspect the payload the IdP
was sending, and that was pretty much impossible.

One can add logging in resource handlers, but if incomplete data
is sent, the resource handler code is not hit, and teh response
is served back by the Server handlers.

Logging the payload (for patch/put/create) with Debugf will
greatly help.